requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
chromadb>=0.5.0
watchdog>=4.0.0
python-frontmatter>=1.1.0
requests>=2.31.0
pypdf>=4.0.0

